+++ Assets/Scripts/Value/Statistics.cs
@@ -1,10 +1,10 @@
 public class Statistics
 {
-    public int Latest { get; set; }
+    public float Latest { get; set; }
 
     public float Avg { get; set; }
 
-    public int Max { get; set; }
+    public float Max { get; set; }
 
-    public int Min { get; set; }
+    public float Min { get; set; }
 }
